+++ Simple Game Library/META-INF/CHANGES.md
@@ -10,6 +10,7 @@ Date: *2016-09-??*
 - Add `setTitle` method to `Graphics` interface.
 - Added option to force graphics mode in `GfxConfigurator'.
 - Added resource loader from POWIDL (thanks a lot to Quantum Reboot).
+- Added hasInputListener method to input interface.
 
 ## Version 1.4.1
 Date: *2016-06-06*